truck: API 정리
