+++ client/src/api/order/getDownFormOrders.ts
@@ -1,5 +1,6 @@
 import { API_END_POINT } from "@/constant";
 import { httpClient } from "@/api/axios";
+import { useQuery } from "@tanstack/react-query";
 
 export const getDownFormOrders = async ({
   limit = 2000,
@@ -13,3 +14,11 @@ export const getDownFormOrders = async ({
   const response = await httpClient.get(API_END_POINT.DOWN_FORM_ORDERS, { params, timeout: 100000 });
   return response.data;
 };
+
+
+export const useOrderList = () => {
+  return useQuery({
+    queryKey: ['downFormOrders'],
+    queryFn: () => getDownFormOrders({ limit: 2000 }),
+  });
+};
