+++ client/src/api/order/index.ts
@@ -6,4 +6,12 @@ export * from './postExcelToDb';
 export * from './postBulkFilter';
 export * from './postBulkWithoutFilter';
 export * from './getBatchInfoAll';
-export * from './getBatchInfoLatest';
+export * from './getDownFormOrders';
+export * from './postBulkDownFormOrders';
+export * from './putBlukDownFormOrders';
+export * from './deleteBulkDownFormOrders';
+export * from './deleteAll';
+export * from './deleteDuplicate';
+export * from './getDownFormOrdersPagination';
+export * from './postExcelToMinio';
+export * from './postExcelToDb';
