+++ client/src/api/order/postExcelToMinio.ts
@@ -2,6 +2,7 @@ import { API_END_POINT } from "@/constant";
 import { httpClient } from "@/api/axios";
 import type { SimpleExcelUploadRequest, ExcelUploadResponse } from "@/api/types";
 
+// 미사용
 export const postExcelToMinio =  async (data: SimpleExcelUploadRequest) => {
   const formData = new FormData();
