+++ client/src/hooks/orderManagement/index.ts
@@ -1,6 +1,4 @@
 export * from "./useOrderData";
 export * from './useOrderCreate';
 export * from './useOrderUpdate';
-export * from './useOrderDelete';
-export * from './useOrderList';
-export * from './useBatchInfo'; 
+export * from './useOrderGridActions';
