--- client/src/hooks/orderManagement/useOrderList.ts
@@ -1,9 +1,0 @@
-import { getDownFormOrders } from "@/api/order/getDownFormOrders";
-import { useQuery } from "@tanstack/react-query";
-
-export const useOrderList = () => {
-  return useQuery({
-    queryKey: ['downFormOrders'],
-    queryFn: () => getDownFormOrders({ limit: 2000 }),
-  });
-};
