+++ client/src/hooks/orderManagement/useOrderUpdate.ts
@@ -1,21 +1,9 @@
-import { putBlukDownFormOrders } from "@/api/order/putBlukDownFormOrders";
-import { useMutation, useQueryClient, type UseMutationResult } from "@tanstack/react-query";
+import { type UseMutationResult } from "@tanstack/react-query";
 import type { BulkUpdateOrderItem, BulkUpdateRequest } from "@/api/types";
 import { toast } from "sonner";
 import type { GridApi } from "ag-grid-community";
 
-export const useOrderUpdate = () => {
-  const queryClient = useQueryClient();
-  return useMutation({
-    mutationFn: (req: BulkUpdateRequest) => putBlukDownFormOrders(req),
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ["orders"] });
-    },
-    onError: (error) => {
-      console.error("대량 수정 실패:", error);
-    },
-  });
-}; 
+
 
 export const handleOrderUpdate = async (
   changedRows: BulkUpdateOrderItem[],
